--- a/crates/easter/src/decl.rs
+++ b/crates/easter/src/decl.rs
@@ -22,7 +22,8 @@ pub enum Export {
 
 #[derive(Debug, PartialEq, Clone, TrackingRef, TrackingMut, Untrack)]
 pub enum Decl {
-    Fun(Fun)
+    Fun(Fun),
+    Let(Option<Span>, Vec<Dtor>, Semi)
 }
 
 #[derive(Debug, PartialEq, Clone, TrackingRef, TrackingMut, Untrack)]