--- a/crates/joker/src/lib.rs
+++ b/crates/joker/src/lib.rs
@@ -6,7 +6,6 @@ pub mod token;
 pub mod lexer;
 mod char;
 mod reader;
-mod lookahead;
 mod test;
 pub mod track;
 pub mod error;