Adding support for `let` statement and fixing `for (let in ...)` where `let` is just an identifier reference in `for`.