@RReverser 2016-12-24
Regarding your performance concerns regarding multitoken lookahead storage, now that #28 is in place, got chance to run benchmarks before and after the changes in this PR:

**before**
```
test tests\esprima\test\3rdparty\angular-1.2.5.js       ... bench: 116,389,654 ns/iter (+/- 3,472,645)
test tests\esprima\test\3rdparty\backbone-1.1.0.js      ... bench:  17,211,617 ns/iter (+/- 338,473)
test tests\esprima\test\3rdparty\jquery-1.9.1.js        ... bench:  91,322,225 ns/iter (+/- 1,436,480)
test tests\esprima\test\3rdparty\jquery.mobile-1.4.2.js ... bench: 140,914,199 ns/iter (+/- 1,807,319)
test tests\esprima\test\3rdparty\mootools-1.4.5.js      ... bench:  75,426,765 ns/iter (+/- 811,637)
test tests\esprima\test\3rdparty\underscore-1.5.2.js    ... bench:  14,762,397 ns/iter (+/- 310,507)
test tests\esprima\test\3rdparty\yui-3.12.0.js          ... bench:  71,796,329 ns/iter (+/- 2,568,568)
```

**after**
```
test tests\esprima\test\3rdparty\angular-1.2.5.js       ... bench: 116,040,869 ns/iter (+/- 4,283,454)
test tests\esprima\test\3rdparty\backbone-1.1.0.js      ... bench:  17,169,866 ns/iter (+/- 420,062)
test tests\esprima\test\3rdparty\jquery-1.9.1.js        ... bench:  91,508,400 ns/iter (+/- 2,122,417)
test tests\esprima\test\3rdparty\jquery.mobile-1.4.2.js ... bench: 140,440,448 ns/iter (+/- 2,290,176)
test tests\esprima\test\3rdparty\mootools-1.4.5.js      ... bench:  75,315,799 ns/iter (+/- 638,664)
test tests\esprima\test\3rdparty\underscore-1.5.2.js    ... bench:  14,707,838 ns/iter (+/- 548,381)
test tests\esprima\test\3rdparty\yui-3.12.0.js          ... bench:  72,104,760 ns/iter (+/- 29,209,443)
```

I would say nothing changes performance-wise.