@RReverser 2016-12-24
No difference on Mac either:

**before**
```
test tests/esprima/test/3rdparty/angular-1.2.5.js       ... bench:  61,004,806 ns/iter (+/- 2,631,685)
test tests/esprima/test/3rdparty/backbone-1.1.0.js      ... bench:   8,442,300 ns/iter (+/- 703,404)
test tests/esprima/test/3rdparty/jquery-1.9.1.js        ... bench:  45,774,492 ns/iter (+/- 1,213,995)
test tests/esprima/test/3rdparty/jquery.mobile-1.4.2.js ... bench:  71,431,129 ns/iter (+/- 2,378,926)
test tests/esprima/test/3rdparty/mootools-1.4.5.js      ... bench:  36,972,970 ns/iter (+/- 869,350)
test tests/esprima/test/3rdparty/underscore-1.5.2.js    ... bench:   7,292,344 ns/iter (+/- 723,298)
test tests/esprima/test/3rdparty/yui-3.12.0.js          ... bench:  37,620,773 ns/iter (+/- 1,240,705)
```

**after**
```
test tests/esprima/test/3rdparty/angular-1.2.5.js       ... bench:  60,934,756 ns/iter (+/- 2,315,062)
test tests/esprima/test/3rdparty/backbone-1.1.0.js      ... bench:   8,477,995 ns/iter (+/- 670,845)
test tests/esprima/test/3rdparty/jquery-1.9.1.js        ... bench:  45,752,413 ns/iter (+/- 1,267,981)
test tests/esprima/test/3rdparty/jquery.mobile-1.4.2.js ... bench:  71,252,654 ns/iter (+/- 2,219,761)
test tests/esprima/test/3rdparty/mootools-1.4.5.js      ... bench:  37,147,282 ns/iter (+/- 1,123,428)
test tests/esprima/test/3rdparty/underscore-1.5.2.js    ... bench:   7,254,767 ns/iter (+/- 628,381)
test tests/esprima/test/3rdparty/yui-3.12.0.js          ... bench:  37,071,472 ns/iter (+/- 944,890)
```