@RReverser 2016-12-25
Regarding the other concern ("we should probably prove that 4 is sufficient 😄") - in fact, even 2 is sufficient here and I added `debug_assert` to both character and token lookaheads to be able to prove it during testing and prevent accidental reallocs of lookahead buffers.