@RReverser 2017-01-14
Fixed and simplified few more things, latest CI comparison:

```
Performance comparison:
 name                                                master ns/iter  PR_26 ns/iter  diff ns/iter  diff % 
 tests/esprima/test/3rdparty/angular-1.2.5.js        99,822,181      99,812,522           -9,659  -0.01% 
 tests/esprima/test/3rdparty/backbone-1.1.0.js       13,509,672      13,130,814         -378,858  -2.80% 
 tests/esprima/test/3rdparty/jquery-1.9.1.js         78,761,757      77,963,404         -798,353  -1.01% 
 tests/esprima/test/3rdparty/jquery.mobile-1.4.2.js  121,671,453     118,947,808      -2,723,645  -2.24% 
 tests/esprima/test/3rdparty/mootools-1.4.5.js       65,039,127      63,086,321       -1,952,806  -3.00% 
 tests/esprima/test/3rdparty/underscore-1.5.2.js     12,106,661      11,510,273         -596,388  -4.93% 
 tests/esprima/test/3rdparty/yui-3.12.0.js           61,677,199      59,197,408       -2,479,791  -4.02% 
```