@RReverser 2017-01-14
Given that performance concern is not confirmed by benchmarks (instead, seeing quite stable subtle improvement) and that code became rather simpler while fixes few more bugs / adds another feature, I'm going to merge this.